\name{RBGL-defunct}
\alias{RBGL-defunct}
\title{Defunct Functions in Package \pkg{RBGL}}
%------ PLEASE: put \alias{.} here for  EACH !
\alias{prim.minST}
%
\description{
  The functions or variables listed here are no longer part of the RBGL package.
}
\usage{
prim.minST()
}
\value{ none }
\seealso{
  \code{\link{Defunct}}
}

