\name{FileDep}
\alias{FileDep}
\non_function{}
\title{ FileDep: a graf object representing a file dependency dataset
example in boost graph library }
\usage{data(FileDep)}
\format{
  The format is:
 list()
  attr(*, "vertices")= chr [1:15] "dax\_h" "yow\_h" "boz\_h" "zow\_h" ...
  attr(*, "edges")=List of 19
  ..- attr(*, "costs")= symbol NULL
  ..- attr(*, "class")= chr "edgeStruct"
 - attr(*, "accL")= list()
 - attr(*, "adjL")= list()
 - attr(*, "class")= chr "graf"
}
\source{
Boost Graph Library 
}
\references{
}
